+++ lib/widgets/collection/collection_grid.dart
@@ -25,7 +25,9 @@ import 'package:aves/widgets/common/extensions/media_query.dart';
 import 'package:aves/widgets/common/grid/draggable_thumb_label.dart';
 import 'package:aves/widgets/common/grid/item_tracker.dart';
 import 'package:aves/widgets/common/grid/scaling.dart';
-import 'package:aves/widgets/common/grid/section_layout.dart';
+import 'package:aves/widgets/common/grid/sections/fixed/scale_grid.dart';
+import 'package:aves/widgets/common/grid/sections/list_layout.dart';
+import 'package:aves/widgets/common/grid/sections/section_layout.dart';
 import 'package:aves/widgets/common/grid/selector.dart';
 import 'package:aves/widgets/common/grid/sliver.dart';
 import 'package:aves/widgets/common/grid/theme.dart';
@@ -34,6 +36,7 @@ import 'package:aves/widgets/common/identity/empty.dart';
 import 'package:aves/widgets/common/identity/scroll_thumb.dart';
 import 'package:aves/widgets/common/providers/tile_extent_controller_provider.dart';
 import 'package:aves/widgets/common/thumbnail/decorated.dart';
+import 'package:aves/widgets/common/thumbnail/image.dart';
 import 'package:aves/widgets/common/tile_extent_controller.dart';
 import 'package:aves/widgets/navigation/nav_bar/nav_bar.dart';
 import 'package:flutter/gestures.dart';
@@ -50,7 +53,8 @@ class CollectionGrid extends StatefulWidget {
   static const int columnCountDefault = 4;
   static const double extentMin = 46;
   static const double extentMax = 300;
-  static const double spacing = 2;
+  static const double fixedExtentLayoutSpacing = 2;
+  static const double mosaicLayoutSpacing = 4;
 
   const CollectionGrid({
     super.key,
@@ -64,6 +68,8 @@ class CollectionGrid extends StatefulWidget {
 class _CollectionGridState extends State<CollectionGrid> {
   TileExtentController? _tileExtentController;
 
+  String get settingsRouteKey => widget.settingsRouteKey;
+
   @override
   void dispose() {
     _tileExtentController?.dispose();
@@ -72,14 +78,17 @@ class _CollectionGridState extends State<CollectionGrid> {
 
   @override
   Widget build(BuildContext context) {
-    _tileExtentController ??= TileExtentController(
-      settingsRouteKey: widget.settingsRouteKey,
-      columnCountDefault: CollectionGrid.columnCountDefault,
-      extentMin: CollectionGrid.extentMin,
-      extentMax: CollectionGrid.extentMax,
-      spacing: CollectionGrid.spacing,
-      horizontalPadding: 2,
-    );
+    final spacing = context.select<Settings, double>((s) => s.getTileLayout(settingsRouteKey) == TileLayout.mosaic ? CollectionGrid.mosaicLayoutSpacing : CollectionGrid.fixedExtentLayoutSpacing);
+    if (_tileExtentController?.spacing != spacing) {
+      _tileExtentController = TileExtentController(
+        settingsRouteKey: settingsRouteKey,
+        columnCountDefault: CollectionGrid.columnCountDefault,
+        extentMin: CollectionGrid.extentMin,
+        extentMax: CollectionGrid.extentMax,
+        spacing: spacing,
+        horizontalPadding: 2,
+      );
+    }
     return TileExtentControllerProvider(
       controller: _tileExtentController!,
       child: _CollectionGridContent(),
@@ -260,12 +269,13 @@ class _CollectionScaler extends StatelessWidget {
     final metrics = context.select<TileExtentController, Tuple2<double, double>>((v) => Tuple2(v.spacing, v.horizontalPadding));
     final tileSpacing = metrics.item1;
     final horizontalPadding = metrics.item2;
+    final brightness = Theme.of(context).brightness;
     return GridScaleGestureDetector<AvesEntry>(
       scrollableKey: scrollableKey,
       tileLayout: tileLayout,
       heightForWidth: (width) => width,
       gridBuilder: (center, tileSize, child) => CustomPaint(
-        painter: GridPainter(
+        painter: FixedExtentGridPainter(
           tileLayout: tileLayout,
           tileCenter: center,
           tileSize: tileSize,
@@ -278,12 +288,21 @@ class _CollectionScaler extends StatelessWidget {
         ),
         child: child,
       ),
-      scaledBuilder: (entry, tileSize) => EntryListDetailsTheme(
+      scaledItemBuilder: (entry, tileSize) => EntryListDetailsTheme(
         extent: tileSize.height,
         child: Tile(
           entry: entry,
           thumbnailExtent: context.read<TileExtentController>().effectiveExtentMax,
           tileLayout: tileLayout,
+        ),
+      ),
+      mosaicItemBuilder: (index, targetExtent) => DecoratedBox(
+        decoration: BoxDecoration(
+          color: ThumbnailImage.computeLoadingBackgroundColor(index * 10, brightness).withOpacity(.9),
+          border: Border.all(
+            color: DecoratedThumbnail.borderColor,
+            width: DecoratedThumbnail.borderWidth,
+          ),
         ),
       ),
       child: child,